+++ src/main/java/webserver/loader/FileResourceLoader.java
@@ -2,7 +2,6 @@ package webserver.loader;
 
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
-import webserver.RequestHandler;
 
 import java.io.*;
 import java.util.List;
